+++ modules/boulder/src/test/ts/atomic/api/ValueSchemaRawTest.ts
@@ -4,9 +4,9 @@ import { Fun, Result } from '@ephox/katamari';
 import { KAssert } from '@ephox/katamari-assertions';
 import * as FieldPresence from 'ephox/boulder/api/FieldPresence';
 import * as FieldSchema from 'ephox/boulder/api/FieldSchema';
-import { Processor } from 'ephox/boulder/api/Main';
 import * as Objects from 'ephox/boulder/api/Objects';
 import * as ValueSchema from 'ephox/boulder/api/ValueSchema';
+import { Processor } from 'ephox/boulder/core/ValueProcessor';
 
 UnitTest.test('ValueSchemaRawTest', () => {
   const checkErr = (label: string, expectedPart: string, input: any, processor: Processor) => {