+++ modules/tinymce/src/plugins/table/test/ts/browser/ui/TableClassListButtonsTest.ts
@@ -1,14 +1,14 @@
-import { Assertions, Keys, Waiter } from '@ephox/agar';
 import { afterEach, describe, it } from '@ephox/bedrock-client';
 import { Arr } from '@ephox/katamari';
-import { TinyHooks, TinySelections, TinyUiActions } from '@ephox/mcagar';
-import { SugarBody, SugarElement } from '@ephox/sugar';
+import { TinyHooks } from '@ephox/mcagar';
+import { SugarBody } from '@ephox/sugar';
 import { assert } from 'chai';
 import Editor from 'tinymce/core/api/Editor';
 import { EditorEvent } from 'tinymce/core/api/util/EventDispatcher';
 import { TableModifiedEvent } from 'tinymce/plugins/table/api/Events';
 import Plugin from 'tinymce/plugins/table/Plugin';
 import Theme from 'tinymce/themes/silver/Theme';
+import { pAssertMenuPresence, pAssertNoCheckmarksInMenu, setEditorContentTableAndSelection } from '../../module/test/TableModifiersTestUtils';
 
 describe('browser.tinymce.plugins.table.ui.TableClassListButtonsTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
@@ -40,44 +40,6 @@ describe('browser.tinymce.plugins.table.ui.TableClassListButtonsTest', () => {
     events = [];
   });
 
-  const setEditorContentTableAndSelection = (editor: Editor) => {
-    editor.setContent(
-      '<table>' +
-        '<tbody>' +
-          '<tr>' +
-            '<td>Filler</td>' +
-          '</tr>' +
-        '</tbody>' +
-      '</table>'
-    );
-
-    TinySelections.setSelection(editor, [ 0, 0, 0 ], 0, [ 0, 0, 0 ], 1);
-  };
-
-  const openMenu = (editor: Editor) =>
-    TinyUiActions.clickOnToolbar(editor, 'button[title="Table styles"]');
-
-  const closeMenu = (editor) =>
-    TinyUiActions.keydown(editor, Keys.escape());
-
-  const pCheckMenuPresence = async (editor: Editor, label: string, expected: Record<string, number>, container: SugarElement<HTMLElement>) => {
-    openMenu(editor);
-    await Waiter.pTryUntil('Ensure the correct values are present', () =>
-      Assertions.assertPresence(label, expected, container)
-    );
-    closeMenu(editor);
-  };
-
-  const pNoCheckmarksInMenu = async (editor: Editor, container: SugarElement<HTMLElement>) => {
-    const expected = {
-      '.tox-menu': 1,
-      '.tox-collection__item[aria-checked="true"]': 0,
-      '.tox-collection__item[aria-checked="false"]': 2
-    };
-
-    await pCheckMenuPresence(editor, 'Menu should open, but not have any checkmarks', expected, container);
-  };
-
   const toggleClasses = (editor: Editor, classes: string[]) => {
     Arr.each(classes, (clazz) =>
       editor.execCommand('mceTableToggleClass', false, clazz)
@@ -90,8 +52,8 @@ describe('browser.tinymce.plugins.table.ui.TableClassListButtonsTest', () => {
     const editor = hook.editor();
     const sugarContainer = SugarBody.body();
 
-    setEditorContentTableAndSelection(editor);
-    await pNoCheckmarksInMenu(editor, sugarContainer);
+    setEditorContentTableAndSelection(editor, 1, 1);
+    await pAssertNoCheckmarksInMenu(editor, 'Table styles', 2, sugarContainer);
     toggleClasses(editor, [ 'a' ]);
 
     const expected = {
@@ -99,15 +61,15 @@ describe('browser.tinymce.plugins.table.ui.TableClassListButtonsTest', () => {
       '.tox-collection__item[aria-checked="true"]': 1,
       '.tox-collection__item[aria-checked="false"]': 1
     };
-    await pCheckMenuPresence(editor, 'There should be a checkmark', expected, sugarContainer);
+    await pAssertMenuPresence(editor, 'There should be a checkmark', 'Table styles', expected, sugarContainer);
   });
 
   it('TINY-7476: Ensure that the checkmark appears for two classes', async () => {
     const editor = hook.editor();
     const sugarContainer = SugarBody.body();
 
-    setEditorContentTableAndSelection(editor);
-    await pNoCheckmarksInMenu(editor, sugarContainer);
+    setEditorContentTableAndSelection(editor, 1, 1);
+    await pAssertNoCheckmarksInMenu(editor, 'Table styles', 2, sugarContainer);
     toggleClasses(editor, [ 'a', 'b' ]);
 
     const expected = {
@@ -116,6 +78,6 @@ describe('browser.tinymce.plugins.table.ui.TableClassListButtonsTest', () => {
       '.tox-collection__item[aria-checked="false"]': 0
     };
 
-    await pCheckMenuPresence(editor, 'There should be two checkmarks', expected, sugarContainer);
+    await pAssertMenuPresence(editor, 'There should be two checkmarks', 'Table styles', expected, sugarContainer);
   });
 });